refactor filter dropdown
